# Conflicts:
#	src/main/java/com/springboot/MiniProject/controller/AdminController.java
#	src/main/java/com/springboot/MiniProject/controller/UserController.java
#	src/main/java/com/springboot/MiniProject/serivce/RefreshTokenService.java

+++ src/main/java/com/springboot/MiniProject/repository/FiliereRepository.java
@@ -3,6 +3,6 @@ package com.springboot.MiniProject.repository;
 import com.springboot.MiniProject.entity.Filiere;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FiliereRepository extends JpaRepository<Filiere,Long> {
+public interface FiliereRepository extends JpaRepository<Filiere,Integer> {
     Filiere findByNom(String nom);
 }
